Split validation responsibilities
The validator struct had too many responsibilities and needed to be
refactored. This should increase reusability of the packages and
increase testability.

Two areas for improvement:

The `Reflector` interface for the `SchemaGenerator` has a hard dependency
on `jsonschema` package which isunfortunate when all it needs is to
marshal the returned structure to json.

The `Actor` in the `validate` package leaks the `schema` package in the
return type of the `ValidateMicroservice` method.